Finalprojectdatasetandproposals.md
The Urban Low-income Neighborhoods and transportation dataset focused on 25 urban neighborhoods and found the times and distance to the nearest Whole Foods store. The forms of transportation included walking, bike, car and bus and was calculated through Google Maps. An additional pivot table breaks down the neighborhoods based on region and shows which low-income region is struggling with transportation. The other two proposals analyze monthly bus costs per neighborhood and an estimated amount of convenient stores in each neighborhood.